para subir, aun falta

+++ bank_c/src/main/java/sd/grupo1/server/dto/AccountUserDTO.java
@@ -26,7 +26,7 @@ public interface AccountUserDTO extends Serializable {
      *
      * @return El saldo de la cuenta.
      */
-    int getBalanceAccount();
+    double getBalanceAccount();
 
     /**
      * Obtiene el nombre del usuario asociado con la cuenta.
